pyproject.toml
[project]
name = "geopulse"
version = "0.1.0"
description = "Open source signals tracker - Flight, Satellites, Traffic, Shipping"
authors = [
    {name = "GeoPulse Team"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
